+++ xdsl/interactive/get_all_available_passes.py
@@ -9,7 +9,6 @@ from xdsl.interactive.rewrites import get_all_possible_rewrites
 from xdsl.ir import Dialect
 from xdsl.parser import Parser
 from xdsl.passes import ModulePass, PassPipeline
-from xdsl.pattern_rewriter import RewritePattern
 
 
 def get_available_pass_list(
@@ -18,7 +17,6 @@ def get_available_pass_list(
     input_text: str,
     pass_pipeline: tuple[ModulePass, ...],
     condense_mode: bool,
-    rewrite_by_names_dict: dict[str, dict[str, RewritePattern]],
 ) -> tuple[AvailablePass, ...]:
     """
     This function returns the available pass list file based on an input text string, pass_pipeline and condense_mode.
@@ -33,10 +31,7 @@ def get_available_pass_list(
     if condense_mode:
         pass_list = get_condensed_pass_list(current_module, all_passes)
         # get all individual rewrites
-        individual_rewrites = get_all_possible_rewrites(
-            current_module,
-            rewrite_by_names_dict,
-        )
+        individual_rewrites = get_all_possible_rewrites(current_module)
         pass_list += tuple(individual_rewrites)
     else:
         pass_list = tuple(AvailablePass(p) for _, p in all_passes)
