+++ xdsl/interactive/rewrites.py
@@ -2,15 +2,12 @@ from collections.abc import Sequence
 
 from xdsl.dialects.builtin import ModuleOp
 from xdsl.interactive.passes import AvailablePass
-from xdsl.pattern_rewriter import PatternRewriter, RewritePattern
+from xdsl.pattern_rewriter import PatternRewriter
 from xdsl.traits import HasCanonicalizationPatternsTrait
 from xdsl.transforms import individual_rewrite
 
 
-def get_all_possible_rewrites(
-    module: ModuleOp,
-    rewrite_by_name: dict[str, dict[str, RewritePattern]],
-) -> Sequence[AvailablePass]:
+def get_all_possible_rewrites(module: ModuleOp) -> Sequence[AvailablePass]:
     """
     Function that takes a sequence of IndividualRewrite Patterns and a ModuleOp, and
     returns the possible rewrites.
@@ -20,13 +17,13 @@ def get_all_possible_rewrites(
     res: list[AvailablePass] = []
 
     for op_idx, matched_op in enumerate(module.walk()):
-        pattern_by_name = rewrite_by_name.get(matched_op.name, {}).copy()
+        if (trait := matched_op.get_trait(HasCanonicalizationPatternsTrait)) is None:
+            continue
 
-        if (
-            trait := matched_op.get_trait(HasCanonicalizationPatternsTrait)
-        ) is not None:
-            for pattern in trait.get_canonicalization_patterns():
-                pattern_by_name[type(pattern).__name__] = pattern
+        pattern_by_name = {
+            type(pattern).__name__: pattern
+            for pattern in trait.get_canonicalization_patterns()
+        }
 
         for pattern_name, pattern in pattern_by_name.items():
             cloned_op = tuple(module.clone().walk())[op_idx]
